tests/e2e/requirements.txt
# Core testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0
pytest-xdist>=3.0.0  # For parallel execution

# HTTP client and stubbing
httpx>=0.24.0
responses>=0.23.0
aioresponses>=0.7.0

# MCP client dependencies  
fastmcp>=0.9.0
pydantic>=2.0.0

# Playwright for visual testing
playwright>=1.40.0

# Utility libraries
python-dotenv>=1.0.0
jinja2>=3.1.0  # For test data templates
faker>=19.0.0  # For generating test data
uuid>=1.0.0

# Development and debugging
rich>=13.0.0  # For better test output
pillow>=10.0.0  # For screenshot processing